register column instead of patching
